Trigger parm callbacks on updates
Hi Guillaume! 
Here's a little fix for setting the parm data. If you have a parm that relies on changes callback we have to explicitly call it by hou.Parm.pressButton(). 
For my workflow this one is crucial, not sure though how critical it is.

ps. thanks for this extension btw :) I'm using it for almost a year now and can't imagine working without it